more stepCount related stuff
Now, the stepCount or whatever it is called in RandomBattle.cpp will only increase when the player has actually moved to a different tile and not when you keep trying to move into a tree or something. Also, you can't talk to a mob when you are entering a battle bruv.